+++ babel.config.cjs
@@ -1,0 +1,26 @@
+'use strict';
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+module.exports = function config (api)
+{
+	api.cache(false);
+
+	return {
+
+		presets :
+		[
+			'@babel/env'
+		],
+
+		plugins :
+		[
+
+		],
+
+		ignore :
+		[
+			/(node_modules)/
+		]
+	};
+};